+++ CHANGELOG.rst
@@ -199,12 +199,18 @@ cause some minor breakage or changes (though in many cases no action will be nee
 
 - Changed timestamps in process monitor output, writers, performance reporter and similar places from UTC to local time. 
   This means these timestamps will match up with the times in run.log output which have always been local time. 
-
 - Since `BaseTest.startProcess` now logs stderr/out automatically before aborting, if you previously wrote extensions 
   that manually log stderr/out after process failures (in a try...except/finally block), you may wish to remove them 
   to avoid duplication, or change them to use the new ``onError=`` mechanism. 
 - `pysys.process.common.CommonProcessWrapper.wait` now returns an error if the specified timeout is isn't a positive 
   number (giving the same behaviour as `BaseTest.waitProcess`). 
+- Changed the implementation of the outcome constants such as `pysys.constants.FAILED` to be an instance of class 
+  `pysys.constants.Outcome` rather than an integer. It is unlikely this change will affect existing code (unless you 
+  have created any custom outcome types, which is not documented). The use of objects to represent outcomes allows for 
+  simpler and more efficient conversion to display name using a ``%s`` format string or ``str()`` without the need for 
+  the LOOKUP dictionary (which still works, but is now deprecated). It also allows easier checking if an outcome 
+  represents a failure using `pysys.constants.Outcome.isFailure()`. The `pysys.constants.PRECEDENT` constant is 
+  deprecated in favour of `pysys.constants.OUTCOMES` which has an identical value.
 - Changed the log messages at the end of a test run to say "THERE WERE NO FAILURES" instead of 
   "THERE WERE NO NON PASSES", and similarly for the "Summary of non passes:".